+++ app/Http/Requests/RoomTypeRequest.php
@@ -17,6 +17,8 @@ class RoomTypeRequest extends FormRequest
             'max_duration' => ['present', 'nullable', 'numeric', 'min:1'],
             'max_participants' => ['present', 'nullable', 'numeric', 'min:1'],
             'server_pool' => ['required', 'integer', 'exists:App\Models\ServerPool,id'],
+            'create_meeting_plugin_config' => ['nullable', 'string', 'json', 'max:65000'],
+            'join_meeting_plugin_config' => ['nullable', 'string', 'json', 'max:65000'],
             'restrict' => ['required', 'boolean'],
             'roles' => [Rule::requiredIf($this->boolean('restrict')), 'array'],
             'roles.*' => ['distinct', 'integer', 'exists:App\Models\Role,id'],