+++ database/migrations/2024_05_22_101114_add_plugin_config_to_room_types_table.php
@@ -1,0 +1,30 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class extends Migration
+{
+    /**
+     * Run the migrations.
+     */
+    public function up(): void
+    {
+        Schema::table('room_types', function (Blueprint $table) {
+            $table->text('create_meeting_plugin_config')->nullable();
+            $table->text('join_meeting_plugin_config')->nullable();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     */
+    public function down(): void
+    {
+        Schema::table('room_types', function (Blueprint $table) {
+            $table->dropColumn(['create_meeting_plugin_config', 'join_meeting_plugin_config']);
+
+        });
+    }
+};